requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django
django-admin-rangefilter==0.12.4
django-celery-beat==2.6.0
django-environ==0.11.2
django-safedelete==1.3.3
django-timezone-field==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
kombu==5.3.7
openpyxl==3.1.5
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
gunicorn
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.9
django-admin-rangefilter==0.12.4
django-celery-beat==2.6.0
django-environ==0.11.2
django-safedelete==1.3.3
django-timezone-field==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
kombu==5.3.7
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
